added getColorByChatColor()
