+++ NovaCore-Spigot/src/main/java/net/zeeraa/novacore/spigot/utils/RGBColorData.java
@@ -1,6 +1,7 @@
 package net.zeeraa.novacore.spigot.utils;
 
 import org.bukkit.ChatColor;
+import org.bukkit.Color;
 import org.json.JSONObject;
 
 public class RGBColorData {
@@ -90,6 +91,45 @@ public class RGBColorData {
 
 	public java.awt.Color tooAWTColor() {
 		return new java.awt.Color(r, g, b);
+	}
+
+	public Color getColorByChatColor(net.md_5.bungee.api.ChatColor color) {
+		switch (color) {
+			case BLACK:
+				return Color.BLACK;
+			case DARK_BLUE:
+				return Color.NAVY;
+			case DARK_GREEN:
+				return Color.GREEN;
+			case DARK_AQUA:
+				return Color.TEAL;
+			case DARK_RED:
+				return Color.MAROON;
+			case DARK_PURPLE:
+				return Color.PURPLE;
+			case GOLD:
+				return Color.ORANGE;
+			case GRAY:
+				return Color.SILVER;
+			case DARK_GRAY:
+				return Color.GRAY;
+			case BLUE:
+				return Color.BLUE;
+			case GREEN:
+				return Color.LIME;
+			case AQUA:
+				return Color.AQUA;
+			case RED:
+				return Color.RED;
+			case LIGHT_PURPLE:
+				return Color.FUCHSIA;
+			case YELLOW:
+				return Color.YELLOW;
+			case WHITE:
+				return Color.WHITE;
+			default:
+				return Color.WHITE;
+		}
 	}
 
 	public static RGBColorData fromChatColor(ChatColor color) {
